WIP move TaxonomyParser to TS
WIP remove old tax parser
